feat(api): add patch and delete request handler through id as parameter
