+++ app/routes/api/v1/warehouse/orders/by-id/index.ts
@@ -3,6 +3,6 @@ import { LoaderFunction } from "@remix-run/node";
 export const loader: LoaderFunction = async () => {
   throw new Response(null, {
     status: 400,
-    statusText: "Order Id Not Provided",
+    statusText: "Bad Request",
   });
 };
